code is now c++0x compliant
so I can now compile without -stdlib=libc++ which I think means means we can compile without setting the mac target to 10.9 which I think means we're good to go with the same compilation that anaconda is using!

of course, I'm not sure I nailed everything here, so if things start breaking, this is a good place to look.

would be nice to have unit tests to make sure.  but things are getting a LOT cleaner.  unit tests will come.