Add methods to Network for HDF5 save-load
Had to modify the network_from_pandas_hdf5 function to take
a class as an argument in order to avoid a circular import
with the network module. Now to use the function directly users
will have to do:

    network_from_pandas_hdf5(Network, filename)

But I expect users will mostly use the Network.from_hdf5 method.