can now return the labels of the pois rather than just the distances
just pass include_poi_ids when getting nearest pois, and instead of just distances, you will get the labels of the poi that's at that distance.  So for instance, if you're trying ot access the nearest hospitals, you won't just get the distance to the nearest N hospitals, but also the names of the nearest N hospitals.  As you might expect with a Pandas-centric view, the names of the hospitals are taken from the index of the x-y series that are passed in to init_pois.