wrote some tests to conform this solves the issue
also ended up having to sort the distance-node pairs properly that come out of ch.  for some reason ch gives the top N number of POIs, but not necessarily in the right order, so we have to sort by distance.  In the case that we want the POI ids of the N closest POIs, we neeed to sort the node_ids by distance, so we use std::pair, which is painful but is par for the course for C++.  anyway it gave the right result both before and after