Migrating from exceptions to presence test
Given the general style of the existing codebase, I migrated my use of
exceptions to presence test of elements in containers.

Still, it would be a good idea to define a behavior when the methods
are called with a non-existent element. The first step for that, it
would be to define a behavior for that in the python side.