Solved 3-day long issue and removed initialization code
The `map::insert` call in `libch.cpp` was being called with an existing
network, with POIs defined. The behavior of the method in that case is
to do nothing, because the key is already defined. Now, the element is
checked for presence and in case it exists it is removed and then
replaced.

I removed the creation of an array of POI data structures from the
code. The initPOIs method is still present because it defines two
values (max_distance and max_items).

@fscottfoti We need to define if those values are actually useful, and
in that case if we should pass when creating the network, or at other
initialization step to avoid having to call this particular method.