Optimized string comparisons
Part of the modifications from 0.3 to 0.4 included modifying the `decay` parameter from an `enum` to `std::string`. That is a good modification in terms of functionality but it seems to be that it made the library slower.

In this commit the comparisons are avoided by choosing a decay function before executing the slow loops. In my local environment, the performance improved (for the SEMCOG example) from 48 to 27 seconds. Still, it seems to be far from the 0.3 performance, so more profiling will be necessary to discover other issues.